Version 0.2

+++ mem.c
@@ -26,7 +26,7 @@ uint8_t mem_read(mem_t *mem, uint16_t address)
 
 void mem_read_area(mem_t *mem, uint16_t address, uint8_t data[], size_t size)
 {
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     data[i] = mem->ram[address + i];
   }
 }
@@ -42,7 +42,7 @@ void mem_write(mem_t *mem, uint16_t address, uint8_t value)
 
 void mem_write_area(mem_t *mem, uint16_t address, uint8_t data[], size_t size)
 {
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     mem->ram[address + i] = data[i];
   }
 }
